refactor(test): Refactor GraphService tests (#2944)
