+++ metadata-io/src/test/java/com/linkedin/metadata/graph/ElasticSearchGraphServiceTest.java
@@ -1,19 +1,11 @@
 package com.linkedin.metadata.graph;
 
 import com.linkedin.common.urn.Urn;
-
 import com.linkedin.metadata.graph.elastic.ESGraphQueryDAO;
 import com.linkedin.metadata.graph.elastic.ESGraphWriteDAO;
 import com.linkedin.metadata.graph.elastic.ElasticSearchGraphService;
-import com.linkedin.metadata.query.RelationshipDirection;
-import com.linkedin.metadata.query.RelationshipFilter;
 import com.linkedin.metadata.utils.elasticsearch.IndexConvention;
 import com.linkedin.metadata.utils.elasticsearch.IndexConventionImpl;
-import java.net.URISyntaxException;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.concurrent.TimeUnit;
-import javax.annotation.Nonnull;
 import org.apache.http.HttpHost;
 import org.apache.http.impl.nio.reactor.IOReactorConfig;
 import org.elasticsearch.client.RestClient;
@@ -23,13 +15,12 @@ import org.testcontainers.elasticsearch.ElasticsearchContainer;
 import org.testng.annotations.AfterTest;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.BeforeTest;
-import org.testng.annotations.Test;
 
-import static com.linkedin.metadata.dao.utils.QueryUtils.*;
-import static org.testng.Assert.*;
+import javax.annotation.Nonnull;
+import java.net.URISyntaxException;
+import java.util.concurrent.TimeUnit;
 
-
-public class ElasticSearchGraphServiceTest {
+public class ElasticSearchGraphServiceTest extends GraphServiceTestBase {
 
   private ElasticsearchContainer _elasticsearchContainer;
   private RestHighLevelClient _searchClient;
@@ -80,108 +71,13 @@ public class ElasticSearchGraphServiceTest {
     _elasticsearchContainer.stop();
   }
 
-  @Test
-  public void testAddEdge() throws Exception {
-    Edge edge1 = new Edge(
-        Urn.createFromString("urn:li:dataset:(urn:li:dataPlatform:kafka,SampleKafkaDataset,PROD)"),
-        Urn.createFromString("urn:li:dataset:(urn:li:dataPlatform:hive,SampleHiveDataset,PROD)"),
-        "DownstreamOf");
-
-    _client.addEdge(edge1);
-    TimeUnit.SECONDS.sleep(5);
-
-    List<String> edgeTypes = new ArrayList<>();
-    edgeTypes.add("DownstreamOf");
-    RelationshipFilter relationshipFilter = new RelationshipFilter();
-    relationshipFilter.setDirection(RelationshipDirection.OUTGOING);
-    relationshipFilter.setCriteria(EMPTY_FILTER.getCriteria());
-
-    List<String> relatedUrns = _client.findRelatedUrns(
-        "",
-        newFilter("urn", "urn:li:dataset:(urn:li:dataPlatform:kafka,SampleKafkaDataset,PROD)"),
-        "",
-        EMPTY_FILTER,
-        edgeTypes,
-        relationshipFilter,
-        0,
-        10);
-
-    assertEquals(relatedUrns.size(), 1);
+  @Override
+  protected @Nonnull GraphService getGraphService() {
+    return _client;
   }
 
-  @Test
-  public void testAddEdgeReverse() throws Exception {
-    Edge edge1 = new Edge(
-        Urn.createFromString("urn:li:dataset:(urn:li:dataPlatform:hive,SampleHiveDataset,PROD)"),
-        Urn.createFromString("urn:li:dataset:(urn:li:dataPlatform:kafka,SampleKafkaDataset,PROD)"),
-        "DownstreamOf");
-
-    _client.addEdge(edge1);
+  @Override
+  protected void syncAfterWrite() throws Exception {
     TimeUnit.SECONDS.sleep(5);
-
-    List<String> edgeTypes = new ArrayList<>();
-    edgeTypes.add("DownstreamOf");
-    RelationshipFilter relationshipFilter = new RelationshipFilter();
-    relationshipFilter.setDirection(RelationshipDirection.INCOMING);
-    relationshipFilter.setCriteria(EMPTY_FILTER.getCriteria());
-
-    List<String> relatedUrns = _client.findRelatedUrns(
-        "",
-        newFilter("urn", "urn:li:dataset:(urn:li:dataPlatform:kafka,SampleKafkaDataset,PROD)"),
-        "",
-        EMPTY_FILTER,
-        edgeTypes,
-        relationshipFilter,
-        0,
-        10);
-
-    assertEquals(relatedUrns.size(), 1);
-  }
-
-  @Test
-  public void testRemoveEdgesFromNode() throws Exception {
-    Edge edge1 = new Edge(
-        Urn.createFromString("urn:li:dataset:(urn:li:dataPlatform:hive,SampleHiveDataset,PROD)"),
-        Urn.createFromString("urn:li:dataset:(urn:li:dataPlatform:kafka,SampleKafkaDataset,PROD)"),
-        "DownstreamOf");
-
-    _client.addEdge(edge1);
-    TimeUnit.SECONDS.sleep(5);
-
-    List<String> edgeTypes = new ArrayList<>();
-    edgeTypes.add("DownstreamOf");
-    RelationshipFilter relationshipFilter = new RelationshipFilter();
-    relationshipFilter.setDirection(RelationshipDirection.INCOMING);
-    relationshipFilter.setCriteria(EMPTY_FILTER.getCriteria());
-
-    List<String> relatedUrns = _client.findRelatedUrns(
-        "",
-        newFilter("urn", "urn:li:dataset:(urn:li:dataPlatform:kafka,SampleKafkaDataset,PROD)"),
-        "",
-        EMPTY_FILTER,
-        edgeTypes,
-        relationshipFilter,
-        0,
-        10);
-
-    assertEquals(relatedUrns.size(), 1);
-
-    _client.removeEdgesFromNode(Urn.createFromString(
-        "urn:li:dataset:(urn:li:dataPlatform:kafka,SampleKafkaDataset,PROD)"),
-        edgeTypes,
-        relationshipFilter);
-    TimeUnit.SECONDS.sleep(5);
-
-    List<String> relatedUrnsPostDelete = _client.findRelatedUrns(
-        "",
-        newFilter("urn", "urn:li:dataset:(urn:li:dataPlatform:kafka,SampleKafkaDataset,PROD)"),
-        "",
-        EMPTY_FILTER,
-        edgeTypes,
-        relationshipFilter,
-        0,
-        10);
-
-    assertEquals(relatedUrnsPostDelete.size(), 0);
   }
 }
