+++ metadata-io/src/test/java/com/linkedin/metadata/graph/Neo4jGraphServiceTest.java
@@ -1,21 +1,14 @@
 package com.linkedin.metadata.graph;
 
-import com.linkedin.common.urn.Urn;
-import com.linkedin.metadata.query.RelationshipDirection;
-import com.linkedin.metadata.query.RelationshipFilter;
-import java.util.ArrayList;
-import java.util.List;
 import org.neo4j.driver.Driver;
 import org.neo4j.driver.GraphDatabase;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
-import org.testng.annotations.Test;
 
-import static com.linkedin.metadata.dao.utils.QueryUtils.*;
-import static org.testng.Assert.*;
+import javax.annotation.Nonnull;
 
 
-public class Neo4jGraphServiceTest {
+public class Neo4jGraphServiceTest extends GraphServiceTestBase {
 
   private Neo4jTestServerBuilder _serverBuilder;
   private Driver _driver;
@@ -34,104 +27,12 @@ public class Neo4jGraphServiceTest {
     _serverBuilder.shutdown();
   }
 
-  @Test
-  public void testAddEdge() throws Exception {
-    Edge edge1 = new Edge(
-        Urn.createFromString("urn:li:dataset:(urn:li:dataPlatform:kafka,SampleKafkaDataset,PROD)"),
-        Urn.createFromString("urn:li:dataset:(urn:li:dataPlatform:hive,SampleHiveDataset,PROD)"),
-        "DownstreamOf");
-
-    _client.addEdge(edge1);
-
-    List<String> edgeTypes = new ArrayList<>();
-    edgeTypes.add("DownstreamOf");
-    RelationshipFilter relationshipFilter = new RelationshipFilter();
-    relationshipFilter.setDirection(RelationshipDirection.OUTGOING);
-    relationshipFilter.setCriteria(EMPTY_FILTER.getCriteria());
-
-    List<String> relatedUrns = _client.findRelatedUrns(
-        "",
-        newFilter("urn", "urn:li:dataset:(urn:li:dataPlatform:kafka,SampleKafkaDataset,PROD)"),
-        "",
-        EMPTY_FILTER,
-        edgeTypes,
-        relationshipFilter,
-        0,
-        10);
-
-    assertEquals(relatedUrns.size(), 1);
+  @Override
+  protected @Nonnull GraphService getGraphService() {
+    return _client;
   }
 
-  @Test
-  public void testAddEdgeReverse() throws Exception {
-    Edge edge1 = new Edge(
-        Urn.createFromString("urn:li:dataset:(urn:li:dataPlatform:hive,SampleHiveDataset,PROD)"),
-        Urn.createFromString("urn:li:dataset:(urn:li:dataPlatform:kafka,SampleKafkaDataset,PROD)"),
-        "DownstreamOf");
+  @Override
+  protected void syncAfterWrite() { }
 
-    _client.addEdge(edge1);
-
-    List<String> edgeTypes = new ArrayList<>();
-    edgeTypes.add("DownstreamOf");
-    RelationshipFilter relationshipFilter = new RelationshipFilter();
-    relationshipFilter.setDirection(RelationshipDirection.INCOMING);
-    relationshipFilter.setCriteria(EMPTY_FILTER.getCriteria());
-
-    List<String> relatedUrns = _client.findRelatedUrns(
-        "",
-        newFilter("urn", "urn:li:dataset:(urn:li:dataPlatform:kafka,SampleKafkaDataset,PROD)"),
-        "",
-        EMPTY_FILTER,
-        edgeTypes,
-        relationshipFilter,
-        0,
-        10);
-
-    assertEquals(relatedUrns.size(), 1);
-  }
-
-  @Test
-  public void testRemoveEdgesFromNode() throws Exception {
-    Edge edge1 = new Edge(
-        Urn.createFromString("urn:li:dataset:(urn:li:dataPlatform:hive,SampleHiveDataset,PROD)"),
-        Urn.createFromString("urn:li:dataset:(urn:li:dataPlatform:kafka,SampleKafkaDataset,PROD)"),
-        "DownstreamOf");
-
-    _client.addEdge(edge1);
-
-    List<String> edgeTypes = new ArrayList<>();
-    edgeTypes.add("DownstreamOf");
-    RelationshipFilter relationshipFilter = new RelationshipFilter();
-    relationshipFilter.setDirection(RelationshipDirection.INCOMING);
-    relationshipFilter.setCriteria(EMPTY_FILTER.getCriteria());
-
-    List<String> relatedUrns = _client.findRelatedUrns(
-        "",
-        newFilter("urn", "urn:li:dataset:(urn:li:dataPlatform:kafka,SampleKafkaDataset,PROD)"),
-        "",
-        EMPTY_FILTER,
-        edgeTypes,
-        relationshipFilter,
-        0,
-        10);
-
-    assertEquals(relatedUrns.size(), 1);
-
-    _client.removeEdgesFromNode(Urn.createFromString(
-        "urn:li:dataset:(urn:li:dataPlatform:kafka,SampleKafkaDataset,PROD)"),
-        edgeTypes,
-        relationshipFilter);
-
-    List<String> relatedUrnsPostDelete = _client.findRelatedUrns(
-        "",
-        newFilter("urn", "urn:li:dataset:(urn:li:dataPlatform:kafka,SampleKafkaDataset,PROD)"),
-        "",
-        EMPTY_FILTER,
-        edgeTypes,
-        relationshipFilter,
-        0,
-        10);
-
-    assertEquals(relatedUrnsPostDelete.size(), 0);
-  }
 }
